# api/openapi/marketplace.v1.yaml
openapi: 3.1.0
info:
  title: Marketplace API
  version: 1.0.0
  description: Публичный REST API маркетплейса. Контракт является источником правды (contract-first).
servers:
  - url: https://api.marketplace.kz
  - url: http://localhost:8080
...
paths:
...
  /api/v1/catalog/categories/{id}:
    get:
      tags: [catalog]
      operationId: getCategory
      parameters:
        - $ref: "#/components/parameters/PathID"
      responses:
        "200":
          description: Категория с путём и эффективной схемой атрибутов (собственные плюс унаследованные)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/catalog/products/{id}:
    get:
      tags: [catalog]
      operationId: getProduct
      description: Опубликованная карточка доступна всем; черновики и карточки на модерации — владельцу и модератору.
      parameters:
        - $ref: "#/components/parameters/PathID"
      responses:
        "200":
          description: Карточка товара
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/catalog/products/{id}/offers:
    get:
      tags: [catalog]
      operationId: listProductOffers
      description: Активные офферы продавцов с правом продажи, отсортированные по цене.
      parameters:
        - $ref: "#/components/parameters/PathID"
      responses:
        "200":
          description: Офферы товара
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Offer" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/admin/catalog/categories:
    post:
      tags: [catalog-admin]
      operationId: createCategory
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                parent_id: { type: string, format: uuid }
                name: { type: string, maxLength: 100 }
                slug: { type: string, maxLength: 100 }
      responses:
        "201":
          description: Категория создана
          content:
            application/json:
              schema:
                type: object
                required: [category_id]
                properties:
                  category_id: { type: string, format: uuid }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /api/v1/admin/catalog/categories/{id}:
    put:
      tags: [catalog-admin]
      operationId: renameCategory
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name: { type: string }
                slug: { type: string }
      responses:
        "204": { description: Категория переименована }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/admin/catalog/categories/{id}/attributes/{code}:
    put:
      tags: [catalog-admin]
      operationId: defineCategoryAttribute
      description: "Код атрибута уникален в пределах ветки дерева: конфликт с предком или потомком — 409."
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
        - $ref: "#/components/parameters/AttributeCode"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, type]
              properties:
                name: { type: string }
                type: { type: string, enum: [string, number, boolean, enum, unit] }
                required: { type: boolean }
                filterable: { type: boolean }
                options:
                  type: array
                  items: { type: string }
                unit: { type: string }
      responses:
        "204": { description: Атрибут определён }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [catalog-admin]
      operationId: removeCategoryAttribute
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
        - $ref: "#/components/parameters/AttributeCode"
      responses:
        "204": { description: Атрибут удалён }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/admin/catalog/moderation-queue:
    get:
      tags: [catalog-admin]
      operationId: listModerationQueue
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Очередь карточек на модерации, старые первыми
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductPage" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/admin/catalog/products/{id}:
    get:
      tags: [catalog-admin]
      operationId: getProductForModeration
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
      responses:
        "200":
          description: Карточка товара с оригиналами изображений
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/admin/catalog/products/{id}/publish:
    post:
      tags: [catalog-admin]
      operationId: publishProduct
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "204": { description: Карточка опубликована }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /api/v1/admin/catalog/products/{id}/reject:
    post:
      tags: [catalog-admin]
      operationId: rejectProduct
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        "204": { description: Карточка отклонена }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /api/v1/seller/sellers/{seller_id}/products:
    post:
      tags: [seller-catalog]
      operationId: createProduct
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/SellerPathID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductContent" }
      responses:
        "201":
          description: Карточка создана в статусе draft
          content:
            application/json:
              schema:
                type: object
                required: [product_id]
                properties:
                  product_id: { type: string, format: uuid }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
    get:
      tags: [seller-catalog]
      operationId: listSellerProducts
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/SellerPathID"
        - { name: status, in: query, schema: { type: string, enum: [draft, on_moderation, published, rejected] } }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Карточки продавца, новые первыми
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/seller/products/{id}:
    get:
      tags: [seller-catalog]
      operationId: getSellerProduct
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
      responses:
        "200":
          description: Карточка продавца
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [seller-catalog]
      operationId: updateProduct
      description: Редактирование доступно в статусах draft и rejected; после правки отклонённая карточка возвращается в draft.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductContent" }
      responses:
        "204": { description: Карточка обновлена }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /api/v1/seller/products/{id}/submit:
    post:
      tags: [seller-catalog]
      operationId: submitProduct
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "204": { description: Карточка отправлена на модерацию }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /api/v1/seller/products/{id}/images:
    post:
      tags: [seller-catalog]
      operationId: requestImageUpload
      description: Возвращает presigned PUT; после загрузки файла вызывается подтверждение.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content_type, size]
              properties:
                content_type: { type: string, enum: ["image/jpeg", "image/png", "image/webp"] }
                size: { type: integer, format: int64, maximum: 10485760 }
      responses:
        "201":
          description: Ссылка для загрузки
          content:
            application/json:
              schema:
                type: object
                required: [image_id, upload]
                properties:
                  image_id: { type: string, format: uuid }
                  upload: { $ref: "#/components/schemas/UploadTarget" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /api/v1/seller/products/{id}/images/{image_id}/confirm:
    post:
      tags: [seller-catalog]
      operationId: confirmImageUpload
      description: Проверяет размер и тип файла по содержимому; превью генерирует worker.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
        - $ref: "#/components/parameters/ImageID"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "204": { description: Загрузка подтверждена }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /api/v1/seller/products/{id}/images/{image_id}:
    delete:
      tags: [seller-catalog]
      operationId: removeProductImage
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
        - $ref: "#/components/parameters/ImageID"
      responses:
        "204": { description: Изображение удалено }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /api/v1/seller/products/{id}/images/order:
    put:
      tags: [seller-catalog]
      operationId: reorderProductImages
      description: Список должен содержать все изображения карточки ровно по одному разу; первое обработанное становится обложкой.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image_ids]
              properties:
                image_ids:
                  type: array
                  items: { type: string, format: uuid }
      responses:
        "204": { description: Порядок изображений обновлён }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/seller/sellers/{seller_id}/variant-groups:
    post:
      tags: [seller-catalog]
      operationId: createVariantGroup
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/SellerPathID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category_id, axes]
              properties:
                category_id: { type: string, format: uuid }
                axes:
                  type: array
                  minItems: 1
                  maxItems: 3
                  items: { type: string }
      responses:
        "201":
          description: Группа вариантов создана
          content:
            application/json:
              schema:
                type: object
                required: [group_id]
                properties:
                  group_id: { type: string, format: uuid }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/seller/variant-groups/{id}/products/{product_id}:
    delete:
      tags: [seller-catalog]
      operationId: removeVariantMember
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/PathID"
        - $ref: "#/components/parameters/ProductPathID"
      responses:
        "204": { description: Карточка исключена из группы }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
...
    PathID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    BadRequest:
      description: Некорректный запрос
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    Forbidden:
      description: Операция запрещена
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Ресурс не найден
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    Category:
      type: object
      required: [id, name, slug, path, attributes, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        parent_id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        path:
          type: array
          items: { $ref: "#/components/schemas/CategoryNode" }
        attributes:
          type: array
          items: { $ref: "#/components/schemas/Attribute" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...